voice-agent/requirements.txt
sounddevice>=0.4.6
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
langchain>=0.1.0
pvporcupine>=3.0.0
pygame>=2.5.2  # Replaces playsound
